requirements.txt
anthropic>=0.39.0
langgraph>=0.2.0
langchain>=0.3.0
langchain-anthropic>=0.3.0
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.24.0
typing-extensions>=4.9.0
rich
click
flask
flask-socketio
jinja2
waitress 